Fix error re units of s(r)
Fix code so that units of density are converted to electrons per cubic
Angstrom whereas previously they were left as electrons per supercell.

This fixes a problem where local softness was in units of eV^-1 per
supercell.

Update documentation to reflect this change.

Read in volume of supercell in input file.