Fix bug with density units and write jmol files
Fix bug where the density units were left uncorrected (which meant
the density was in units of electrons per supercell)

Now this has been fixed, the local softness is in units of
inverse electron volts per cubic Angstrom.

Previously it was in units of inverse eV per supercell, which
couldn't be compared so well between different supercells and
therefore different metal surfaces.

Make two new files for use with jmol

One is the den_fmt file for zero charge in units of electrons per
cubic Angstrom. This file has zeros in the lower half of the cell
for easier visualisation.

The other new file is a file of the negative of the local softness
which is in units of inverse electron volts per cubic Angstrom.
As for the density file, the lower half of the cell is artificially
set to zero. Also make certain points in the slab region artificially
high in local softness. This will allow the use of sed to make a
softness file with a specific minimum and maximum so that the range
for all the colourmaps can be made the same