Improve location of min and max for setting range
Read in minimum and maximum value for setting range from input file
which will save time when plotting.

Fix problem where minimum and maximum values for setting range are
visible in the region where I will be making graphics. This looked
strange.

Instead of locating the minimum and maximum values for setting the
range in the region which will be in the final graphic region
displayed, locate them between 2/3 and the bottom of the unit cell.

Also make the density and softness zero between 1/2 and 2/3 so that
the unit cell will look neater. When making final graphics, I'll be
able to just crop off the bottom of the unit cell and it should look
better than before.