Fix density units for Bader topology.data input file
The units previously outputted for the density in the topology.data
input file were in electrons per cubic Angstrom. In fact while these
are better units for most uses, the topology code written by Steve
assumes that the density will be in electrons per supercell. Therefore
to avoid a warning the density in electrons per supercell is read in
and outputted in the topology.data output file.