Fix units in output files
There were persistent problems with the units in previous versions
of this code. At first both density and local softness were in units
of per supercell, which is as outputted by CASTEP in the den_fmt
output file. However for visualising the local softness and density
in 3D and for comparing different surfaces, it would be better to have
units of per cubic Angstrom.

Initially when I changed the units I also changed the units of the
topology.data input file to also be per cubic Angstrom. However this
doesn't work properly as the topology code expects units of per
supercell for density and 'spin' aka local softness.

Fix code so that the topology.data file has units of per supercell.
Likewise _s_r.dat file has units of per supercell.

However, make units of s_r.den_fmt and _jmol_s_r.den_fmt in units
of eV^-1 per cubic Angstrom. Also make _halfcell.den_fmt in units
of per cubic Angstrom.

This should mean that the local softness will be in units of eV^-1
A^-3 and the atomic softness will be in units of eV^-1.